fix the random number generation bug

+++ memcached_client/util.h
@@ -20,7 +20,7 @@ void timingTests();
 void timestampTest();
 void lockTest();
 int randomFunction();
-int parRandomFunction(struct worker* worker);
+double parRandomFunction(struct worker* worker);
 double round(double d);
